+++ changelog.md
@@ -1,5 +1,10 @@
 # The Modding Tree changelog:
 
+- Layer updates now have a determined order and starts with earlier-rowed layers.
+- Automation now has a determined order and starts with later-rowed layers.
+- Commas should no longer appear in the decimal places of a number.
+- Fixed potential issue in displaying the tree.
+
 ### v2.1.2 - 10/19/20
 - Added buyUpgrade function (buyUpg still works though)
 - Added author name to modInfo.